# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "koora API",
        "contact": {
            "name": "Ahmed Magdy and abouzaid",
            "url": "https://www.google.com"
        },
        "version": "1.0"
    },
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/Seasons/{season}/Event/matches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "//get all matches in specific season swagger",
                "description": "get all matches in specific season",
                "operationId": "GET_ALL_matches",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "description": "season ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Match"
                ],
                "description": "post Create matche form ",
                "operationId": "POST_Create_matches",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "description": "season ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/match_creation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Seasons/{season}/Event/matches/{date}": {
            "get": {
                "tags": [
                    "Match"
                ],
                "description": "get all matches in specific season in specific date",
                "operationId": "GET_ALL_matches_by_date",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "description": "season ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "description": "date",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Seasons/{season}/Event/{team}/matches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "description": "get all matches for specific team in specific season",
                "operationId": "GET_ALL_matches_for_team",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "description": "season ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "team",
                        "in": "path",
                        "description": "RegisteredTeam id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/matches/{date}": {
            "get": {
                "tags": [
                    "Match"
                ],
                "description": "get all matches in specific date",
                "operationId": "GET_ALL_matches_by_date",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "date",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Players/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "description": "get Players with it's id",
                "operationId": "GET_Player",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Remove player from database",
                "operationId": "deleteplayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "player ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/Seasons/{season_id}/RegisteredTeam/{team_id}/RegisteredPlayer/players": {
            "get": {
                "tags": [
                    "Registered Players"
                ],
                "description": "get all registered Players in a team",
                "operationId": "AllPlayers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "description": "Season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "team_id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "match_creation": {
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2017-12-17"
                },
                "time": {
                    "type": "string",
                    "example": "01:00:00"
                },
                "stage_id": {
                    "type": "integer",
                    "example": 1
                },
                "team_1_id": {
                    "type": "integer",
                    "example": 1
                },
                "team_2_id": {
                    "type": "integer",
                    "example": 2
                },
                "stadium": {
                    "type": "string",
                    "example": "Ahly stadium"
                }
            }
        },
...
    }
}